% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brt_helper_functions.R
\name{brt_cross_val}
\alias{brt_cross_val}
\title{Run a BRT with cross validation to determine the best ntrees and important
variables}
\usage{
brt_cross_val(spp, pred_data, learning_rate = 0.001, ...)
}
\arguments{
\item{spp}{species name}

\item{pred_data}{model data}

\item{learning_rate}{learning rate}

\item{...}{other arguments passed to gbm.step}
}
\description{
Run a BRT with cross validation to determine the best ntrees and important
variables
}
