% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_functions.R
\name{crop_transform}
\alias{crop_transform}
\title{Crop and then transform}
\usage{
crop_transform(ras, bbx, crs_out)
}
\arguments{
\item{ras}{SpatRaster}

\item{bbx}{sf polygon of bbox}

\item{crs_out}{crs to use. As returned by \code{terra::crs}}
}
\value{
cropped and transformed SpatRaster
}
\description{
Transforms bbox to raster CRS, then crops raster, then transforms raster to
given CRS
}
