% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{cut.fn}
\alias{cut.fn}
\title{Rasterize a series of spatial predictions (needed for plotting)}
\usage{
\method{cut}{fn}(
  df = NA,
  target_raster = NA,
  column_name = NA,
  lower_bound = NA,
  upper_bound = NA
)
}
\description{
Rasterize a series of spatial predictions (needed for plotting)
}
