% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_functions.R
\name{do_sp_compare}
\alias{do_sp_compare}
\title{Compare species across models}
\usage{
do_sp_compare(sp6, sp4 = "XXXX", samp_pts, val_dat, in_dat_pth)
}
\arguments{
\item{sp6}{6 letter species common name code}

\item{sp4}{4 letter species common name code}

\item{samp_pts}{SpatVector of points to compare across}

\item{in_dat_pth}{root folder containing all models, will be searched
recursively to match the species code}
}
\value{
Prints species code, maps and comparison graphs
}
\description{
Compare species across models
}
