% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_functions.R
\name{extract_transform}
\alias{extract_transform}
\title{Transform points then extract}
\usage{
extract_transform(ras, pts)
}
\arguments{
\item{ras}{SpatRaster}

\item{pts}{SpatVector of points that will be projected to match the raster
and then used to extract raster values}
}
\value{
data.frame
}
\description{
Transform points then extract
}
