% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{fit_brt}
\alias{fit_brt}
\title{Fit boosted regression tree models}
\usage{
fit_brt(model_data, response_column = NA, covariate_columns = NA)
}
\description{
Fit boosted regression tree models
}
