% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_points_extract.R
\name{gee_points_extract}
\alias{gee_points_extract}
\title{Extract data from GEE}
\usage{
gee_points_extract(
  meth_path,
  loc.yr,
  out_save,
  projection.trans = 5072,
  local.buffer = 200,
  neighbour.buffer = 2000,
  do.test.run = TRUE
)
}
\arguments{
\item{meth_path}{path to table of extraction methods and sources. Will be
updated with download progress. Must match BAM National Models format.
https://docs.google.com/spreadsheets/d/1XATuq8BOYC2KkbJObaturaf4NFD2ufvn/edit?usp=sharing&ouid=104837701987164094932&rtpof=true&sd=true}

\item{loc.yr}{sf object with a unique row id and a year column identifying
each unique site and year}

\item{out_save}{directory where outputs should be saved}

\item{projection.trans}{EPSG code for projection of output}

\item{local.buffer}{Buffer for smaller radius TODO: improve so this is only
set in spreadsheet}

\item{neighbour.buffer}{Buffer for larger radius}

\item{do.test.run}{Should a test run of downloading 20 points from GEE be
attempted first?}
}
\value{
Files are saved to the out_save directory. And the spreadsheet at
meth_path is updated with Complete = 1 and Running = 0 for variables that
were downloaded successfully
}
\description{
Extract data from GEE for buffered point locations for a set of covariates
described in a table
}
