% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brt_helper_functions.R
\name{make_pred_data}
\alias{make_pred_data}
\title{Make dataframe of response and predictors for one species}
\usage{
make_pred_data(spp, samp_row_ind, y, off, pred_vars, pred_var_nms)
}
\arguments{
\item{spp}{species names}

\item{samp_row_ind}{index of rows to use. Based on resampling}

\item{y}{sparse matrix with response variable}

\item{off}{offsets}

\item{pred_vars}{predictor variables}

\item{pred_var_nms}{predictor variable names}
}
\description{
Make dataframe of response and predictors for one species
}
