% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brt_helper_functions.R
\name{plot_fun}
\alias{plot_fun}
\title{Plot marginal effect plots for the top 12 variables}
\usage{
plot_fun(res, rel_inf, spp)
}
\arguments{
\item{res}{model object}

\item{rel_inf}{table of relative influence}

\item{spp}{species name}
}
\description{
Plot marginal effect plots for the top 12 variables
}
