% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROFBirds-package.R
\docType{package}
\name{ROFBirds-package}
\alias{ROFBirds}
\alias{ROFBirds-package}
\title{ROFBirds: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sarah Endicott \email{sarah.endicott@canada.ca} (\href{https://orcid.org/0000-0001-9644-5343}{ORCID})

}
\keyword{internal}
