% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brt_helper_functions.R
\name{run_cross_val}
\alias{run_cross_val}
\title{Do cross validation step}
\usage{
run_cross_val(
  spp,
  samp_row_ind,
  samp_col_ind,
  y,
  off,
  pred_vars,
  pred_var_nms,
  learning_rate = 0.001,
  save_dir,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
make data, fit model, save model, plot response, return ntrees, relative
influence in table
}
