% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{wt_dl_loc_year}
\alias{wt_dl_loc_year}
\title{Download location and year information from WildTrax}
\usage{
wt_dl_loc_year(proj_id, sens_id)
}
\arguments{
\item{proj_id}{proect_id in wildTrax}
}
\description{
Download location and year information from WildTrax
}
